+++ index.html
@@ -8,7 +8,6 @@
 
     <link rel="stylesheet" href="./css/style.css">
     <link rel="stylesheet" href="./css/style.css">
-    <script src="./scr/main.js"></script>
     
 </head>
 
@@ -17,7 +16,7 @@
         <h1>스파르타 영화 평점 콜랙션</h1>
     </header>
 
-    <form class="search" onsubmit="handleSearch(event)">
+    <form class="search">
         <lable>영화검색 : </lable>
         <input  type="text" id="search-input" placeholder="영화를 검색 해주세요!">
         <button type="submit">검색</button>
@@ -28,6 +27,7 @@
             
         </div>
     </main>
+    <script type="module" src="./scr/main.js"></script>
 </body>
 
 </html>
